+++ site/src/scripts/appcastDownload/state.js
@@ -1,0 +1,2 @@
+const appcastCloseTimers = new WeakMap();
+const appcastCloseDelayMs = 140;
